Add symlink file to generate symlinks in user bin
Created a python script that will link all bin scripts within
this directory to $HOME/bin using symlinks. This is to prep for
creating a "new" bin command script that will make the script
in this repo at user_scripts/bin_scripts which will back propogate
to this repo with symlinks.